adding country coordinates from globe to schema itself
